requirements.txt
django==2.2.7
wheel
django-rest-framework==0.1.0
django-cors-headers==3.1.1
django-simple_history==2.7.3
channels-redis==2.4.1
djangorestframework-yaml==1.0.3
redis==3.3.11
pyyaml==5.1.2
psycopg2==2.8.4
drf-yasg==1.17.0
django-filter==2.2.0
paho-mqtt==1.5.0
channels
chasgimqtt